feat: add custom file path to be iterated

+++ cidr-information/CIDR-check.php
@@ -24,7 +24,7 @@ $lockFilePath = __DIR__ . "/tmp/$basename.lock";
 $statusFile = __DIR__ . "/status.txt";
 $short_opts = "p:m::";
 $long_opts = [
-  "proxy:",
+  "path:",
   "max::",
   "userId::",
   "lockFile::",
@@ -49,12 +49,19 @@ if (!empty($options['admin']) && $options['admin'] !== 'false') {
   set_time_limit(0);
 }
 
+$folder = tmp() . '/ips-ports';
+if (!empty($options['path']) && file_exists($options['path'])) {
+  $filePath = $options['path'];
+} else {
+  $filePath = getRandomFileFromFolder($folder, 'txt');
+}
+
 if (file_exists($lockFilePath) && !is_debug() && !$isAdmin) {
   echo date(DATE_RFC3339) . ' another process still running' . PHP_EOL;
   exit();
 } else {
-  file_put_contents($lockFilePath, date(DATE_RFC3339));
-  file_put_contents($statusFile, 'scan generated IP:PORT');
+  write_file($lockFilePath, date(DATE_RFC3339));
+  write_file($statusFile, 'scan generated IP:PORT');
 }
 
 function exitProcess()
@@ -63,7 +70,7 @@ function exitProcess()
   if (file_exists($lockFilePath)) {
     unlink($lockFilePath);
   }
-  file_put_contents($statusFile, 'idle');
+  write_file($statusFile, 'idle');
 }
 
 register_shutdown_function('exitProcess');
@@ -71,45 +78,47 @@ register_shutdown_function('exitProcess');
 // set memory
 ini_set('memory_limit', '2024M');
 
-$folder = tmp() . '/ips-ports';
-$filePath = getRandomFileFromFolder($folder, 'txt');
-$outputPath = __DIR__ . '/../proxies.txt';
+// main script start
+
 $startTime = time();
 $str_to_remove = [];
-if (filesize($filePath) == 0) {
-  unlink($filePath);
-  exit("$filePath size 0");
-}
 
-iterateBigFilesLineByLine([$filePath], $max, function (string $line) use (&$str_to_remove, $startTime, $filePath, $maxExecutionTime) {
-  $db = new ProxyDB();
-  $proxies = extractProxies($line, null, false);
-  for ($i = 0; $i < count($proxies); $i++) {
-    if (!is_debug()) {
-      if (time() - $startTime > $maxExecutionTime) {
-        echo "Execution time exceeded. Stopping execution." . PHP_EOL;
-        break;
-      }
-    }
-    $item = $proxies[$i];
-    if (isPortOpen($item->proxy)) {
-      // add to database on port open
-      $db->updateData($item->proxy, ['status' => 'untested']);
-      //      $http = checkProxy($item->proxy);
-      //      $socks5 = checkProxy($item->proxy, 'socks5');
-      //      $socks4 = checkProxy($item->proxy, 'socks4');
-      //      if ($http || $socks4 || $socks5) {
-      //        echo "$item->proxy working" . PHP_EOL;
-      //      } else {
-      //        echo "$item->proxy port open, but not proxy" . PHP_EOL;
-      //      }
-    } else {
-      echo "$item->proxy port closed" . PHP_EOL;
-    }
-    $str_to_remove[] = $item->proxy;
-    Scheduler::register(function () use ($filePath, $str_to_remove) {
-      // echo 'removing ' . count($str_to_remove) . ' lines' . PHP_EOL;
-      echo removeStringFromFile($filePath, $str_to_remove) . PHP_EOL;
-    }, 'clean-up-' . basename(__FILE__));
+if (file_exists($filePath)) {
+  if (filesize($filePath) == 0) {
+    unlink($filePath);
+    exit("$filePath size 0");
   }
-});
+
+  iterateBigFilesLineByLine([$filePath], $max, function (string $line) use (&$str_to_remove, $startTime, $filePath, $maxExecutionTime) {
+    $db = new ProxyDB();
+    $proxies = extractProxies($line, null, false);
+    for ($i = 0; $i < count($proxies); $i++) {
+      if (!is_debug()) {
+        if (time() - $startTime > $maxExecutionTime) {
+          echo "Execution time exceeded. Stopping execution." . PHP_EOL;
+          break;
+        }
+      }
+      $item = $proxies[$i];
+      if (isPortOpen($item->proxy)) {
+        // add to database on port open
+        $db->updateData($item->proxy, ['status' => 'untested']);
+        //      $http = checkProxy($item->proxy);
+        //      $socks5 = checkProxy($item->proxy, 'socks5');
+        //      $socks4 = checkProxy($item->proxy, 'socks4');
+        //      if ($http || $socks4 || $socks5) {
+        //        echo "$item->proxy working" . PHP_EOL;
+        //      } else {
+        //        echo "$item->proxy port open, but not proxy" . PHP_EOL;
+        //      }
+      } else {
+        echo "$item->proxy port closed" . PHP_EOL;
+      }
+      $str_to_remove[] = $item->proxy;
+      Scheduler::register(function () use ($filePath, $str_to_remove) {
+        // echo 'removing ' . count($str_to_remove) . ' lines' . PHP_EOL;
+        echo removeStringFromFile($filePath, $str_to_remove) . PHP_EOL;
+      }, 'clean-up-' . basename(__FILE__));
+    }
+  });
+}
